Refactor: Signup/Login - handler&useHook&commonModal
